requirements.txt
adal==1.2.7
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
argcomplete==3.6.2
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
autocommand==2.2.2
azure-common==1.1.28
azure-core==1.35.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==13.0.0
azure-mgmt-core==1.6.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-network==29.0.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==23.0.0
azureml-core==1.60.0.post1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
boto3==1.40.19
botocore==1.40.19
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
configobj==5.0.9
contextlib2==21.6.0
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
databricks-sdk==0.64.0
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.24.1
durationpy==0.10
dvc==3.62.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gto==1.8.0
gunicorn==23.0.0
h11==0.16.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jaraco.text==4.0.0
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonpickle==4.1.1
kiwisolver==1.4.9
knack==0.12.0
kombu==5.5.4
kubernetes==33.1.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mlflow==3.3.1
mlflow-skinny==3.3.1
mlflow-tracing==3.3.1
more-itertools==10.7.0
msal==1.33.0
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4.post1
multidict==6.6.4
ndg-httpsclient==0.5.1
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
omegaconf==2.3.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.2
packaging==25.0
pandas==2.3.2
paramiko==3.5.1
pathspec==0.12.1
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.3.8
prometheus_client==0.22.1
prometheus_flask_exporter==0.23.2
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
py4j==0.10.9.7
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pysftp==0.2.9
PySocks==1.7.1
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-auth-aws-sigv4==0.7
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.1
scmrepo==3.5.2
seaborn==0.13.2
SecretStorage==3.3.3
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sqltrie==0.11.2
starlette==0.47.3
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
virtualenv==20.34.0
voluptuous==0.15.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0
